Test bugsnag encoding against schema
Unfortunately there wasn't an official JSON Schema for the report format
that I could find, but there is a lengthy JSON sample from which we can
generate a schema. That schema is probably a bit too strict at the
moment (it assumes every field found it required), so we might need to
tweak it a bit.