Add module for buffering operations
We have an API that we can use to write reports of multiple events, but
we don't offer any help for batching up these events. We'd like to add
an additional API for writing individual events, but in a way that they
end up getting batched into HTTP requests.

This commit sets up the `Buffer` abstraction that will allow us to get
there.