Remove dependency of tests on python package
We had a cobbled-together JSON schema and used it to validate the JSON
this library produces for reports. Unfortunately there don't seem to be
any recent Haskell packages for validating JSON schema's in the
ecosystem and so we were calling out to a python package.

This python dependency in our tests would result in tests not running on
other machines where this dependency is missing. So we simplify the
tests here instead to not require any external dependencies.